+++ solution/WebApplication/WebApplication/Controllers/SourceAndTargetSystemsJsonSchemaController.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore;
 using WebApplication.Services;
-
+using WebApplication.Framework;
 using WebApplication.Models;
 
 namespace WebApplication.Controllers
@@ -17,7 +17,7 @@ namespace WebApplication.Controllers
         protected readonly AdsGoFastContext _context;
         
 
-        public SourceAndTargetSystemsJsonSchemaController(AdsGoFastContext context, SecurityAccessProvider securityAccessProvider) : base(securityAccessProvider)
+        public SourceAndTargetSystemsJsonSchemaController(AdsGoFastContext context, ISecurityAccessProvider securityAccessProvider, IEntityRoleProvider roleProvider) : base(securityAccessProvider, roleProvider)
         {
             Name = "SourceAndTargetSystemsJsonSchema";
             _context = context;
@@ -30,6 +30,7 @@ namespace WebApplication.Controllers
         }
 
         // GET: SourceAndTargetSystemsJsonSchema/Details/5
+        [ChecksUserAccess]
         public async Task<IActionResult> Details(string id)
         {
             if (id == null)
@@ -40,9 +41,10 @@ namespace WebApplication.Controllers
             var sourceAndTargetSystemsJsonSchema = await _context.SourceAndTargetSystemsJsonSchema
                 .FirstOrDefaultAsync(m => m.SystemType == id);
             if (sourceAndTargetSystemsJsonSchema == null)
-            {
                 return NotFound();
-            }
+            if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                return new ForbidResult();
+
 
             return View(sourceAndTargetSystemsJsonSchema);
         }
@@ -59,11 +61,16 @@ namespace WebApplication.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [ChecksUserAccess]
         public async Task<IActionResult> Create([Bind("SystemType,JsonSchema")] SourceAndTargetSystemsJsonSchema sourceAndTargetSystemsJsonSchema)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(sourceAndTargetSystemsJsonSchema);
+                if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                {
+                    return new ForbidResult();
+                }
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(IndexDataTable));
             }
@@ -71,6 +78,7 @@ namespace WebApplication.Controllers
         }
 
         // GET: SourceAndTargetSystemsJsonSchema/Edit/5
+        [ChecksUserAccess()]
         public async Task<IActionResult> Edit(string id)
         {
             if (id == null)
@@ -80,9 +88,10 @@ namespace WebApplication.Controllers
 
             var sourceAndTargetSystemsJsonSchema = await _context.SourceAndTargetSystemsJsonSchema.FindAsync(id);
             if (sourceAndTargetSystemsJsonSchema == null)
-            {
                 return NotFound();
-            }
+
+            if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                return new ForbidResult();
             return View(sourceAndTargetSystemsJsonSchema);
         }
 
@@ -91,6 +100,7 @@ namespace WebApplication.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [ChecksUserAccess]
         public async Task<IActionResult> Edit(string id, [Bind("SystemType,JsonSchema")] SourceAndTargetSystemsJsonSchema sourceAndTargetSystemsJsonSchema)
         {
             if (id != sourceAndTargetSystemsJsonSchema.SystemType)
@@ -103,6 +113,10 @@ namespace WebApplication.Controllers
                 try
                 {
                     _context.Update(sourceAndTargetSystemsJsonSchema);
+
+                    if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                        return new ForbidResult();
+			
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -122,6 +136,7 @@ namespace WebApplication.Controllers
         }
 
         // GET: SourceAndTargetSystemsJsonSchema/Delete/5
+        [ChecksUserAccess]
         public async Task<IActionResult> Delete(string id)
         {
             if (id == null)
@@ -132,19 +147,25 @@ namespace WebApplication.Controllers
             var sourceAndTargetSystemsJsonSchema = await _context.SourceAndTargetSystemsJsonSchema
                 .FirstOrDefaultAsync(m => m.SystemType == id);
             if (sourceAndTargetSystemsJsonSchema == null)
-            {
                 return NotFound();
-            }
+		
+            if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                return new ForbidResult();
 
             return View(sourceAndTargetSystemsJsonSchema);
         }
 
         // POST: SourceAndTargetSystemsJsonSchema/Delete/5
         [HttpPost, ActionName("Delete")]
+        [ChecksUserAccess()]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
             var sourceAndTargetSystemsJsonSchema = await _context.SourceAndTargetSystemsJsonSchema.FindAsync(id);
+
+            if (!await CanPerformCurrentActionOnRecord(sourceAndTargetSystemsJsonSchema))
+                return new ForbidResult();
+		
             _context.SourceAndTargetSystemsJsonSchema.Remove(sourceAndTargetSystemsJsonSchema);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(IndexDataTable));
